Add basic stubs and files to build the new or1k BSP: or1ksim.
This is not meant to be upstreamed, it's just some trials that may
help, in the time being, in debugging bare-metal applications and
minimal
basic infrastructure for or1k architecture. However, it can be
a starting point when moving to RTEMS porting task.
The following are major changes:

+cpukit/cpu: added or1k cpu, some files are borrowed and modified
from the old or32 port. Stubs, definitions and changes made to be
able to build libraries and move to bsp.

+libbsp/or1k: Minimal configuration for adding new or1k/or1ksim
BSP. This bsp only has start.S which do nothing other than calling
main. It's provided to make sure that libbsp.a is generated when
building RTEMS for or1k.

Conflicts:
	cpukit/score/cpu/or1k/Makefile.am
	cpukit/score/cpu/or1k/cpu.c
	cpukit/score/cpu/or1k/rtems/asm.h
	cpukit/score/cpu/or1k/rtems/score/cpu.h
	cpukit/score/cpu/or1k/rtems/score/cpu_asm.h
	cpukit/score/cpu/or1k/rtems/score/or1k.h
	cpukit/score/cpu/or1k/rtems/score/types.h